test: BridgeGameTest 으로 이름 변경, 테스트 추가
